add ProductPolicy

+++ app/Http/Controllers/ProductController.php
@@ -7,11 +7,15 @@ use App\Http\Requests\StoreProductRequest;
 use App\Http\Requests\UpdateProductRequest;
 use App\Http\Resources\ProductResource;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Support\Facades\Gate;
 
 class ProductController extends Controller
 {
     public function index()
     {
+        // Gate
+        Gate::authorize('viewAny', Product::class);
+        
         $data = Product::all();
 
         return response()->json([
@@ -22,6 +26,9 @@ class ProductController extends Controller
 
     public function store(StoreProductRequest $request): JsonResponse
     {
+        // Gate
+        Gate::authorize('create', Product::class);
+
         $newProduct = Product::create($request->validated());
 
         return response()->json([
@@ -33,6 +40,9 @@ class ProductController extends Controller
 
     public function show(Product $product)
     {
+        // Gate
+        Gate::authorize('view', Product::class);
+        
         return response()->json([
             "data" => ProductResource::make($product),
         ]);
@@ -41,6 +51,9 @@ class ProductController extends Controller
 
     public function update(UpdateProductRequest $request, Product $product)
     {
+        // Gate
+        Gate::authorize('update', Product::class);
+        
         $product->update($request->validated());
 
         return response()->json([
@@ -52,6 +65,9 @@ class ProductController extends Controller
 
     public function destroy(Product $product)
     {
+        // Gate
+        Gate::authorize('delete', Product::class);
+        
         $product->delete();
 
         return response()->json([
